Add capabilities to run Newts on Cassandra
Allow users to run and configure OpenNMS Horizon with Newts on Cassandra. Extend the entrypoint script to allow users
to configure the basic Cassandra connection with environment variables and initializing the Newts Keyspace in the
Cassandra cluster. Add documentation in the README with new introduced start options and -n -c.

Additionally ensure OpenNMS Horizon runs the configuration test before doing a startup or upgrade.